* updated for copyright and license
git-svn-id: http://brian-d-foy.svn.sourceforge.net/svnroot/brian-d-foy/trunk@2030 d9f60a02-fc28-0410-9752-bc0d6c4f57a6

+++ rijndael.h
@@ -56,7 +56,7 @@
 
 #if defined(__MINGW32__) && ! defined(_CRYPT_RIJNDAEL_H_TYPES)
 	#define _CRYPT_RIJNDAEL_H_TYPES
-	typedef unsigned long UINT32;
+	typedef unsigned int UINT32;
 	typedef unsigned char UINT8;
 #endif
